Feature/map layer2

+++ libs/commons/src/main/java/com/recom/commons/calculator/ARGBColor.java
@@ -7,7 +7,7 @@ import lombok.NonNull;
 public class ARGBColor {
 
     @NonNull
-    public static Long RGB(
+    public static int RGB(
             final int r,
             final int g,
             final int b
@@ -16,7 +16,7 @@ public class ARGBColor {
     }
 
     @NonNull
-    public static Long ARGB(
+    public static int ARGB(
             final int a,
             final int r,
             final int g,
@@ -25,7 +25,7 @@ public class ARGBColor {
         if (a < 0 || a > 255 || r < 0 || r > 255 || g < 0 || g > 255 || b < 0 || b > 255) {
             throw new IllegalArgumentException("Invalid color component value. Must be between 0 and 255.");
         } else {
-            return (long) ((a << 24) | (r << 16) | (g << 8) | b);
+            return ((a << 24) | (r << 16) | (g << 8) | b);
         }
     }
